refactor(settings): move account components to features
