+++ frontend/app/features/settings/account/components/AvatarUpload.vue
